Retry without refresh mapping if pool is busy
poolboy:transaction/2 doesn't call the transaction fun with 'full'. Poolboy:checkout/3
can return full if called with Block = false, but in poolboy:transaction/2,3, checkout
is called with Block = true.

Instead, if the pool is full (actually empty?), the caller is added to a waiting list
by poolboy and doens't get a reply until a worker returns to the pool. The reply is sent
explicitly using gen_server:reply/2 when a pool worker is checked in back into the pool.
In the meantime, the poolboy:checkout(Pool, true, Timeout) in transaction/3 times out
and an exit exception on the form exit:{timeout,{gen_server,call,_}} is raised.

Normally, you shouldn't catch an exit exception, but in this case it's OK because
poolboy has already caught it, cancelled the waiting for this worker and re-raised the
exception (try-catch-raise) in poolboy:checkout/3.

The exit exception then bubbles out and is caught in eredis_cluster_pool:transaction/2
and converted to {error, no_connection}, which is handled by refreshing the mapping and
then retrying. This commit changes this to {error, pool_busy} and retry without refresh
mapping.