Implement ASK and MOVED redirects
ASK and MOVED redirects for simple commands and pipelines implemented
and tested.

For pipeline commands, redirects are followed only if the result
contains only one kind of identical redirects.

When ASK redirects are followed in a pipeline command, ASKING is added
before every command, except between MULTI and EXEC/DISCARD, as required
by Redis.

When MOVED redirects are followed, an asynchronous refresh of the
slots mapping is started in the background, but the redirected commands
are performed without waiting for the refresh to be completed.

Redirects are not implemented in qmn/1. A TODO comment is added for that.