Add functions for commands on specific Redis node
Add qn/2, allow 2nd arg of transaction/2 to be a pool and add
get_all_pools/0 (as a wrapper for the internal function with
the same name in eredis_cluster_monitor).

This is a generalization of the mechanism used in scan/4. Scan
is reimplemented using qn/2.

Also adds a missing base case for transaction/4 which is a helper
for optimistic_locking_transaction/3.